+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the CommitPulse extension will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.64] - 2026-08-03
+
+### Added
+
+- Add Dead Branches TreeView to sidebar
+- Refactor Dead Branches Report to use commit_history table
+
 ## [0.1.63] - 2026-08-02
 
 ### Fixed